requirements.txt
arrow==1.4.0
asgiref==3.11.1
binaryornot==0.6.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
cookiecutter==2.7.1
Django==6.0.3
django-browser-reload==1.21.0
django-colorfield==0.14.0
django-htmx==1.27.0
django-tailwind==4.4.2
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pillow==12.1.1
psycopg2-binary==2.9.11
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.33.0
rich==14.3.3
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tzdata==2025.3
urllib3==2.6.3
